refactor: moved handlers into hook directory + few improvements
