+++ src/lib/hooks/chat-contact-click-handler.ts
