+++ src/public-api.ts
@@ -23,6 +23,8 @@ export * from './lib/core/stanza';
 export * from './lib/core/subscription';
 export * from './lib/core/translations';
 export * from './lib/directives/links.directive';
+export * from './lib/hooks/chat-contact-click-handler';
+export * from './lib/hooks/file-upload-handler';
 export * from './lib/ngx-chat.module';
 export * from './lib/services/adapters/xmpp/abstract-stanza-builder';
 export * from './lib/services/adapters/xmpp/plugins/abstract-xmpp-plugin';
@@ -47,8 +49,6 @@ export * from './lib/services/adapters/xmpp/xmpp-chat-adapter.service';
 export * from './lib/services/adapters/xmpp/xmpp-chat-connection.service';
 export * from './lib/services/adapters/xmpp/xmpp-client-factory.service';
 export * from './lib/services/chat-background-notification.service';
-export * from './lib/services/chat-contact-click-handler';
-export * from './lib/services/file-upload-handler';
 export * from './lib/services/chat-list-state.service';
 export * from './lib/services/chat-message-list-registry.service';
 export * from './lib/services/chat-service';
